More cleanup and refactoring
