+++ nd4j/nd4j-backends/nd4j-tests/src/test/java/org/nd4j/autodiff/OpValidationSuite.java
@@ -4,11 +4,7 @@ import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
-import org.nd4j.autodiff.opvalidation.LossOpValidation;
-import org.nd4j.autodiff.opvalidation.MiscOpValidation;
-import org.nd4j.autodiff.opvalidation.ReductionOpValidation;
-import org.nd4j.autodiff.opvalidation.TransformOpValidation;
-import org.nd4j.autodiff.opvalidation.ReductionBpOpValidation;
+import org.nd4j.autodiff.opvalidation.*;
 import org.nd4j.autodiff.validation.OpValidation;
 import org.nd4j.linalg.api.buffer.DataBuffer;
 import org.nd4j.linalg.factory.Nd4j;
@@ -37,8 +33,10 @@ import org.nd4j.linalg.function.Function;
 @Suite.SuiteClasses({
         //Note: these will be run as part of the suite only, and will NOT be run again separately
         ReductionBpOpValidation.class,
+        LayerOpValidation.class,
         LossOpValidation.class,
         MiscOpValidation.class,
+        ReductionBpOpValidation.class,
         ReductionOpValidation.class,
         TransformOpValidation.class
 })
