+++ nd4j/nd4j-backends/nd4j-tests/src/test/java/org/nd4j/autodiff/opvalidation/ShapeOpValidation.java
@@ -1,0 +1,640 @@
+package org.nd4j.autodiff.opvalidation;
+
+import lombok.Builder;
+import lombok.Data;
+import lombok.val;
+import org.junit.Test;
+import org.nd4j.autodiff.samediff.SDVariable;
+import org.nd4j.autodiff.samediff.SameDiff;
+import org.nd4j.autodiff.validation.OpValidation;
+import org.nd4j.autodiff.validation.TestCase;
+import org.nd4j.linalg.api.ndarray.INDArray;
+import org.nd4j.linalg.checkutil.NDArrayCreationUtil;
+import org.nd4j.linalg.factory.Nd4j;
+import org.nd4j.linalg.factory.Nd4jBackend;
+import org.nd4j.linalg.primitives.Pair;
+import org.nd4j.linalg.primitives.Triple;
+import org.nd4j.linalg.util.ArrayUtil;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+import static org.junit.Assert.assertEquals;
+import static org.nd4j.linalg.indexing.NDArrayIndex.interval;
+
+public class ShapeOpValidation extends BaseOpValidation {
+    public ShapeOpValidation(Nd4jBackend backend) {
+        super(backend);
+    }
+
+    /*
+    To test:
+    tile
+    reshape
+    permute
+    expandDims
+    repeat
+    rollAxis
+    doRepeat
+     */
+
+    @Test
+    public void testConcat() {
+
+//        int[] concatDim = new int[]{0,0,0,1,1,1,2,2,2};
+        int[] concatDim = new int[]{0, 0, 0};
+        List<List<int[]>> origShapes = new ArrayList<>();
+        origShapes.add(Arrays.asList(new int[]{3, 4}, new int[]{5, 4}));
+        origShapes.add(Arrays.asList(new int[]{1, 2, 3}, new int[]{1, 2, 3}, new int[]{2, 2, 3}));
+        origShapes.add(Arrays.asList(new int[]{1, 2, 3, 4}, new int[]{2, 2, 3, 4}));
+
+        List<String> failed = new ArrayList<>();
+
+        for (int i = 0; i < concatDim.length; i++) {
+
+            SameDiff sd = SameDiff.create();
+            List<int[]> shapes = origShapes.get(i);
+
+            SDVariable[] toConcat = new SDVariable[shapes.size()];
+            INDArray[] orig = new INDArray[shapes.size()];
+            for (int j = 0; j < shapes.size(); j++) {
+                orig[j] = Nd4j.rand(shapes.get(j));
+                toConcat[j] = sd.var("concat-in-" + String.valueOf(j), orig[j]);
+            }
+
+            SDVariable sdConcat = sd.concat("c", 0, toConcat);
+            SDVariable stdev = sd.standardDeviation("out", sdConcat, true);
+
+            String msg = "i=" + i + ", concatDim=" + concatDim[i];
+            TestCase tc = new TestCase(sd);
+            tc.testName(msg)
+                    .expectedOutput("c", Nd4j.concat(concatDim[i], orig));
+
+            String error = OpValidation.validate(tc);
+            if(error != null){
+                failed.add(name);
+            }
+        }
+
+        assertEquals(failed.toString(), 0, failed.size());
+    }
+
+    @Test
+    public void testReshapeGradient() {
+        int[] origShape = new int[]{3, 4, 5};
+
+        List<String> failed = new ArrayList<>();
+
+        for (int[] toShape : new int[][]{{3, 4 * 5}, {3 * 4, 5}, {1, 3 * 4 * 5}, {3 * 4 * 5, 1}}) {
+            for (Pair<INDArray, String> p : NDArrayCreationUtil.getAll3dTestArraysWithShape(12345, origShape)) {
+                INDArray inArr = p.getFirst().muli(100);
+
+                SameDiff sd = SameDiff.create();
+                SDVariable in = sd.var("in", inArr);
+                SDVariable reshape = sd.reshape(in, toShape);
+                //Using stdev here: mean/sum would backprop the same gradient for each input...
+                SDVariable stdev = sd.standardDeviation("out", reshape, true);
+
+                INDArray out = sd.execAndEndResult();
+                INDArray expOut = in.getArr().std(true, Integer.MAX_VALUE);
+
+                String msg = "toShape=" + Arrays.toString(toShape) + ", source=" + p.getSecond();
+                TestCase tc = new TestCase(sd);
+                tc.testName(msg)
+                        .expectedOutput("out", expOut);
+
+                String error = OpValidation.validate(tc);
+                if(error != null){
+                    failed.add(name);
+                }
+            }
+        }
+
+        assertEquals(failed.toString(), 0, failed.size());
+    }
+
+    @Test
+    public void testPermuteGradient() {
+        int[] origShape = new int[]{3, 4, 5};
+
+        List<String> failed = new ArrayList<>();
+
+        for (int[] perm : new int[][]{{0, 1, 2}, {0, 2, 1}, {1, 0, 2}, {1, 2, 0}, {2, 0, 1}, {2, 1, 0}}) {
+            for (Pair<INDArray, String> p : NDArrayCreationUtil.getAll3dTestArraysWithShape(12345, origShape)) {
+                String msg = "permute=" + Arrays.toString(perm) + ", source=" + p.getSecond();
+                System.out.println(msg);
+
+                INDArray inArr = p.getFirst().muli(100);
+
+                SameDiff sd = SameDiff.create();
+                SDVariable in = sd.var("in", inArr);
+                SDVariable permute = sd.f().permute(in, perm);
+                //Using stdev here: mean/sum would backprop the same gradient for each input...
+                SDVariable stdev = sd.standardDeviation("out", permute, true);
+
+                INDArray out = sd.execAndEndResult();
+                INDArray expOut = in.getArr().std(true, Integer.MAX_VALUE);
+                assertEquals(msg, expOut, out);
+
+
+                TestCase tc = new TestCase(sd);
+                tc.testName(msg)
+                        .expectedOutput("out", expOut);
+
+                String error = OpValidation.validate(tc);
+                if(error != null){
+                    failed.add(name);
+                }
+            }
+        }
+
+        assertEquals(failed.toString(), 0, failed.size());
+    }
+
+    @Test
+    public void testExpandDimsGradient() {
+        val origShape = new long[]{3, 4};
+
+        List<String> failed = new ArrayList<>();
+
+        boolean first = true;
+        for (int i = 0; i < 3; i++) {
+
+            long[] expExpandShape;
+            switch (i) {
+                case 0:
+                    expExpandShape = new long[]{1, 3, 4};
+                    break;
+                case 1:
+                    expExpandShape = new long[]{3, 1, 4};
+                    break;
+                case 2:
+                    expExpandShape = new long[]{3, 4, 1};
+                    break;
+                default:
+                    throw new RuntimeException();
+            }
+
+            for (Pair<INDArray, String> p : NDArrayCreationUtil.getAllTestMatricesWithShape(origShape[0], origShape[1], 12345)) {
+                INDArray inArr = p.getFirst().muli(100);
+
+                SameDiff sd = SameDiff.create();
+                SDVariable in = sd.var("in", inArr);
+                SDVariable expand = sd.f().expandDims(in, i);
+                //Using stdev here: mean/sum would backprop the same gradient for each input...
+                SDVariable stdev = sd.standardDeviation("out", expand, true);
+
+                INDArray out = sd.execAndEndResult();
+                INDArray expOut = in.getArr().std(true, Integer.MAX_VALUE);
+
+                assertArrayEquals(expExpandShape, expand.getArr().shape());
+                INDArray expExpand = inArr.dup('c').reshape(expExpandShape);
+
+                String msg = "expandDim=" + i + ", source=" + p.getSecond();
+                log.info("Starting: " + msg);
+
+                TestCase tc = new TestCase(sd);
+                tc.testName(msg)
+                        .expectedOutput("out", expOut)
+                        .expectedOutput(expand.getVarName(), expExpand);
+
+                String error = OpValidation.validate(tc);
+                if(error != null){
+                    failed.add(name);
+                }
+            }
+        }
+        assertEquals(failed.toString(), 0, failed.size());
+    }
+
+    @Test
+    public void testSqueezeGradient() {
+        val origShape = new long[]{3, 4, 5};
+
+        List<String> failed = new ArrayList<>();
+
+        for (int i = 0; i < 3; i++) {
+
+            val shape = origShape.clone();
+            shape[i] = 1;
+
+            for (Pair<INDArray, String> p : NDArrayCreationUtil.getAll3dTestArraysWithShape(12345, shape)) {
+                INDArray inArr = p.getFirst().muli(100);
+
+                SameDiff sd = SameDiff.create();
+                SDVariable in = sd.var("in", inArr);
+                SDVariable squeeze = sd.f().squeeze(in, i);
+                //Using stdev here: mean/sum would backprop the same gradient for each input...
+                SDVariable stdev = sd.standardDeviation("out", squeeze, true);
+
+                long[] expShapePostSqueeze;
+                switch (i) {
+                    case 0:
+                        expShapePostSqueeze = new long[]{4, 5};
+                        break;
+                    case 1:
+                        expShapePostSqueeze = new long[]{3, 5};
+                        break;
+                    case 2:
+                        expShapePostSqueeze = new long[]{3, 4};
+                        break;
+                    default:
+                        throw new RuntimeException();
+                }
+
+                sd.execAndEndResult();
+
+                INDArray squeezed = squeeze.getArr();
+//                assertArrayEquals(expShapePostSqueeze, squeezed.shape());
+
+                INDArray out = sd.execAndEndResult();
+                INDArray expOut = in.getArr().std(true, Integer.MAX_VALUE);
+                assertEquals(expOut, out);
+
+                String msg = "squeezeDim=" + i + ", source=" + p.getSecond();
+                TestCase tc = new TestCase(sd)
+                        .expectedOutput("out", expOut);
+
+
+                String error = OpValidation.validate(tc);
+                if(error != null){
+                    failed.add(name);
+                }
+            }
+        }
+
+        assertEquals(failed.toString(), 0, failed.size());
+    }
+
+
+    @Test
+    public void testSliceGradient() {
+        Nd4j.getRandom().setSeed(12345);
+
+        //Order here: original shape, begin, size
+        List<Triple<int[], int[], int[]>> testCases = new ArrayList<>();
+        testCases.add(new Triple<>(new int[]{3, 4}, new int[]{0, 0}, new int[]{3, 4}));
+        testCases.add(new Triple<>(new int[]{3, 4}, new int[]{1, 1}, new int[]{3, 4}));
+        testCases.add(new Triple<>(new int[]{3, 4}, new int[]{1, 2}, new int[]{2, 3}));
+        testCases.add(new Triple<>(new int[]{3, 4, 5}, new int[]{0, 0, 0}, new int[]{3, 4, 5}));
+        testCases.add(new Triple<>(new int[]{3, 4, 5}, new int[]{1, 1, 1}, new int[]{2, 3, 4}));
+        testCases.add(new Triple<>(new int[]{3, 4, 5}, new int[]{1, 0, 2}, new int[]{3, 3, 4}));
+
+        List<String> failed = new ArrayList<>();
+
+        for (int i = 0; i < testCases.size(); i++) {
+            Triple<int[], int[], int[]> t = testCases.get(i);
+            int[] os = t.getFirst();
+            int[] b = t.getSecond();
+            int[] e = t.getThird();
+            INDArray arr = Nd4j.rand(os);
+
+            SameDiff sd = SameDiff.create();
+            SDVariable in = sd.var("in", arr);
+            SDVariable slice = sd.slice(in, b, e);
+            SDVariable stdev = sd.standardDeviation(slice, true);
+
+            String msg = "i=" + i + ": inShape=" + Arrays.toString(os) + ", begin=" + Arrays.toString(b) + ", end=" + Arrays.toString(e);
+            log.info("Starting test: " + msg);
+
+            TestCase tc = new TestCase(sd);
+            String error = OpValidation.validate(tc);
+            if(error != null){
+                failed.add(name);
+            }
+        }
+
+        assertEquals(failed.toString(), 0, failed.size());
+    }
+
+
+    @Builder(builderClassName = "Builder")
+    @Data
+    private static class SSCase {
+        private int[] shape;
+        private int[] begin;
+        private int[] end;
+        private int[] strides;
+        private int beginMask;
+        private int endMask;
+        private int ellipsisMask;
+        private int newAxisMask;
+        private int shrinkAxisMask;
+
+        public static class Builder {
+
+            public Builder shape(int... shape) {
+                this.shape = shape;
+                return this;
+            }
+
+            public Builder begin(int... begin) {
+                this.begin = begin;
+                return this;
+            }
+
+            public Builder end(int... end) {
+                this.end = end;
+                return this;
+            }
+
+            public Builder strides(int... strides) {
+                this.strides = strides;
+                return this;
+            }
+        }
+    }
+
+    @Test
+    public void testStridedSliceGradient() {
+        Nd4j.getRandom().setSeed(12345);
+
+        //Order here: original shape, begin, size
+        List<SSCase> testCases = new ArrayList<>();
+        testCases.add(SSCase.builder().shape(3, 4).begin(0, 0).end(3, 4).strides(1, 1).build());
+        testCases.add(SSCase.builder().shape(3, 4).begin(1, 1).end(2, 3).strides(1, 1).build());
+        testCases.add(SSCase.builder().shape(3, 4).begin(-999, 0).end(3, 4).strides(1, 1).beginMask(1).build());
+        testCases.add(SSCase.builder().shape(3, 4).begin(1, 1).end(3, -999).strides(1, 1).endMask(1 << 1).build());
+        testCases.add(SSCase.builder().shape(3, 4).begin(-999, 0).end(-999, 4).strides(1, 1).beginMask(1).endMask(1).build());
+        testCases.add(SSCase.builder().shape(3, 4).begin(-999, 0, 0).end(-999, 3, 4).strides(1, 1).newAxisMask(1).build());
+
+        testCases.add(SSCase.builder().shape(3, 4, 5).begin(0, 0, 0).end(3, 4, 5).strides(1, 1, 1).build());
+        testCases.add(SSCase.builder().shape(3, 4, 5).begin(1, 2, 3).end(3, 4, 5).strides(1, 1, 1).build());
+        testCases.add(SSCase.builder().shape(3, 4, 5).begin(0, 0, 0).end(3, 3, 5).strides(1, 2, 2).build());
+        testCases.add(SSCase.builder().shape(3, 4, 5).begin(1, -999, 1).end(3, 3, 4).strides(1, 1, 1).beginMask(1 << 1).build());
+        testCases.add(SSCase.builder().shape(3, 4, 5).begin(1, -999, 1).end(3, 3, -999).strides(1, 1, 1).beginMask(1 << 1).endMask(1 << 2).build());
+        testCases.add(SSCase.builder().shape(3, 4, 5).begin(1, 2).end(3, 4).strides(1, 1).ellipsisMask(1 << 1).build());   //[1:3,...,2:4]
+        testCases.add(SSCase.builder().shape(3, 4, 5).begin(1, -999, 1, 2).end(3, -999, 3, 4).strides(1, -999, 1, 2).newAxisMask(1 << 1).build());
+        testCases.add(SSCase.builder().shape(3, 4, 5).begin(1, 0, 1).end(3, -999, 4).strides(1, 1, 1).shrinkAxisMask(1 << 1).build());
+        testCases.add(SSCase.builder().shape(3, 4, 5).begin(1, 1, 1).end(3, -999, 4).strides(1, 1, 1).shrinkAxisMask(1 << 1).build());
+
+        List<String> failed = new ArrayList<>();
+
+        for (int i = 0; i < testCases.size(); i++) {
+            SSCase t = testCases.get(i);
+            INDArray arr = Nd4j.rand(t.getShape());
+
+            SameDiff sd = SameDiff.create();
+            SDVariable in = sd.var("in", arr);
+            SDVariable slice = sd.stridedSlice(in, t.getBegin(), t.getEnd(), t.getStrides(), t.getBeginMask(),
+                    t.getEndMask(), t.getEllipsisMask(), t.getNewAxisMask(), t.getShrinkAxisMask());
+            SDVariable stdev = sd.standardDeviation(slice, true);
+
+            String msg = "i=" + i + ": " + t;
+            log.info("Starting test: " + msg);
+
+            TestCase tc = new TestCase(sd);
+            String error = OpValidation.validate(tc);
+            if(error != null){
+                failed.add(name);
+            }
+        }
+        assertEquals(failed.toString(), 0, failed.size());
+    }
+
+
+    @Test
+    public void testMerge() {
+        Nd4j.getRandom().setSeed(12345);
+
+        List<String> failed = new ArrayList<>();
+
+        for (int t = 0; t < 3; t++) {
+            for (int numArrays : new int[]{3, 1}) {
+                for (long[] shape : new long[][]{{1}, {3, 4}, {3, 4, 5}}) {
+
+                    SameDiff sd = SameDiff.create();
+                    SDVariable[] arr = new SDVariable[numArrays];
+
+                    for (int i = 0; i < numArrays; i++) {
+                        arr[i] = sd.var(String.valueOf(i), Nd4j.rand(shape));
+                    }
+
+                    SDVariable merge;
+                    switch (t) {
+                        case 0:
+                            merge = sd.mergeAdd(arr);
+                            break;
+                        case 1:
+                            merge = sd.mergeMax(arr);
+                            break;
+                        case 2:
+                            merge = sd.mergeAvg(arr);
+                            break;
+                        default:
+                            throw new RuntimeException();
+                    }
+
+                    String msg = merge.opName() + " - numArrays=" + numArrays + ", shape=" + Arrays.toString(shape);
+                    SDVariable loss = sd.standardDeviation("loss", merge, true);
+
+                    TestCase tc = new TestCase(sd);
+                    String error = OpValidation.validate(tc);
+                    if(error != null){
+                        failed.add(name);
+                    }
+                }
+            }
+        }
+
+        assertEquals(failed.toString(), 0, failed.size());
+    }
+
+    @Test
+    public void testStack() {
+        Nd4j.getRandom().setSeed(12345);
+
+        List<String> failed = new ArrayList<>();
+
+        List<long[]> origShape = Arrays.asList(
+                new long[]{1},
+                new long[]{1, 1},
+                new long[]{3, 4},
+                new long[]{3, 4, 5},
+                new long[]{3, 4, 5, 6}
+        );
+
+        for (long[] shape : origShape) {
+            for (int axis = 0; axis <= shape.length; axis++) {
+                for (int numInputs : new int[]{1, 3}) {
+
+                    long[] expOutShape = new long[shape.length + 1];
+                    int x = 0;
+                    for (int i = 0; i <= shape.length; i++) {
+                        if (i == axis) {
+                            expOutShape[i] = numInputs;
+                        } else {
+                            expOutShape[i] = shape[x++];
+                        }
+                    }
+
+
+                    SameDiff sd = SameDiff.create();
+
+                    SDVariable[] in = new SDVariable[numInputs];
+                    INDArray[] inArr = new INDArray[numInputs];
+                    for (int i = 0; i < numInputs; i++) {
+                        inArr[i] = Nd4j.rand(shape);
+                        in[i] = sd.var(String.valueOf(i), inArr[i]);
+                    }
+
+                    SDVariable stack = sd.stack(axis, in);
+
+                    INDArray out = sd.execAndEndResult();
+                    assertArrayEquals(expOutShape, out.shape());
+
+                    if (ArrayUtil.prodLong(shape) == 1) {
+                        SDVariable loss = sd.sum("loss", stack);
+                    } else {
+                        SDVariable loss = sd.standardDeviation("loss", stack, true);
+                    }
+
+                    String msg = Arrays.toString(shape) + ", axis=" + axis + ", numInputs=" + numInputs;
+
+                    TestCase tc = new TestCase(sd);
+                    String error = OpValidation.validate(tc);
+                    if(error != null){
+                        failed.add(name);
+                    }
+                }
+            }
+        }
+
+        assertEquals(failed.toString(), 0, failed.size());
+    }
+
+
+    @Test
+    public void testUnStack() {
+        Nd4j.getRandom().setSeed(12345);
+
+        List<String> failed = new ArrayList<>();
+
+        List<long[]> unstackedShape = Arrays.asList(
+                new long[]{1},
+                new long[]{1, 1},
+                new long[]{3, 4},
+                new long[]{3, 4, 5},
+                new long[]{3, 4, 5, 6}
+        );
+
+        for (long[] shape : unstackedShape) {
+            for (int axis = 0; axis <= shape.length; axis++) {
+                for (int numInputs : new int[]{1, 3}) {
+
+                    long[] stackedShape = new long[shape.length + 1];
+                    int x = 0;
+                    for (int i = 0; i <= shape.length; i++) {
+                        if (i == axis) {
+                            stackedShape[i] = numInputs;
+                        } else {
+                            stackedShape[i] = shape[x++];
+                        }
+                    }
+
+
+                    SameDiff sd = SameDiff.create();
+                    INDArray in = Nd4j.rand(stackedShape);
+                    SDVariable var = sd.var("var", in);
+
+                    SDVariable[] unstacked = sd.unstack(var, axis, numInputs);
+
+                    //for gradient check, need to combine to single scalar output...
+                    SDVariable merged = sd.mergeAvg(unstacked);
+
+                    if (ArrayUtil.prodLong(stackedShape) == 1) {
+                        SDVariable loss = sd.sum("loss", merged);
+                    } else {
+                        SDVariable loss = sd.standardDeviation("loss", merged, true);
+                    }
+
+                    String msg = "Unstacked shape = " + Arrays.toString(shape) + ", stacked shape = " + Arrays.toString(stackedShape)
+                            + ", axis=" + axis + ", numInputs=" + numInputs;
+
+                    sd.execAndEndResult();
+                    for (SDVariable v : unstacked) {
+                        assertArrayEquals(msg, shape, v.getArr().shape());
+                    }
+
+                    TestCase tc = new TestCase(sd);
+                    String error = OpValidation.validate(tc);
+                    if(error != null){
+                        failed.add(name);
+                    }
+                }
+            }
+        }
+
+        assertEquals(failed.toString(), 0, failed.size());
+    }
+
+    @Test
+    public void testTile() {
+        Nd4j.getRandom().setSeed(12345);
+
+        List<int[]> tileArg = Arrays.asList(
+                new int[]{1},
+                new int[]{5},
+                new int[]{3,4},
+                new int[]{2,3},
+                new int[]{2,3,4}
+        );
+
+        INDArray[] orig = new INDArray[tileArg.size()];
+        orig[0] = Nd4j.valueArrayOf(new long[]{1}, 3.0);
+        orig[1] = Nd4j.valueArrayOf(new long[]{1}, 3.0);
+        orig[2] = Nd4j.valueArrayOf(new long[]{1,1}, 3.0);
+        orig[3] = Nd4j.linspace(1,4,4).reshape('c', 2,2);
+
+        INDArray[] exp = new INDArray[tileArg.size()];
+        exp[0] = Nd4j.trueVector(new double[]{3});
+        exp[1] = Nd4j.trueVector(new double[]{3,3,3,3,3});
+        exp[2] = Nd4j.valueArrayOf(new long[]{3,4}, 3.0);
+        exp[3] = Nd4j.create(2*2, 2*3);
+        for( int i=0; i<2; i++ ){
+            for( int j=0; j<2; j++ ){
+                exp[3].get(interval(2*i,2*(i+1)), interval(2*j,2*(j+1))).assign(orig[3]);
+            }
+        }
+        exp[4] = Nd4j.create(3*2, 4*3, 5*4);
+        for( int i=0; i<2; i++ ){
+            for( int j=0; j<3; j++ ){
+                for( int k=0; k<4; k++ ) {
+                    exp[4].get(interval(3 * i, 3 * (i + 1)), interval(4 * j, 4 * (j + 1)), interval(5*k, 5*(k+1))).assign(orig[4]);
+                }
+            }
+        }
+
+        List<String> failed = new ArrayList<>();
+
+        for (int i = 0; i < tileArg.size(); i++) {
+            int[] tArg = tileArg.get(i);
+            INDArray inArr = orig[i];
+
+            SameDiff sd = SameDiff.create();
+            SDVariable var = sd.var("in", inArr);
+            SDVariable tile = sd.tile(var, tArg);
+
+            if(exp[i].length() == 1){
+                SDVariable loss = sd.sum("loss", tile);
+            } else {
+                SDVariable loss = sd.standardDeviation("loss", tile, true);
+            }
+
+            sd.execAndEndResult();
+            INDArray tiled = tile.getArr();
+            assertEquals(exp[i], tiled);
+
+            String msg = "Shape=" + Arrays.toString(inArr.shape()) + " - tile=" + Arrays.toString(tArg);
+
+            TestCase tc = new TestCase(sd);
+            String error = OpValidation.validate(tc);
+            if(error != null){
+                failed.add(name);
+            }
+        }
+
+        assertEquals(failed.toString(), 0, failed.size());
+    }
+}
